requirements.txt
requests==2.26.0
Jinja2==3.1.2
python-dotenv==0.17.0